# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My First API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/v1/rooms/find/": {
            "get": {
                "summary": "find Room",
                "operationId": "4f0399cb4407d3c2c4c4660d226b496a",
                "responses": {
                    "200": {
                        "description": "An example endpoint"
                    }
                }
            },
            "post": {
                "summary": "create room",
                "operationId": "b9880766f8b8df9bf3227b12fef38180",
                "parameters": [
                    {
                        "parameter": "hotelId",
                        "name": "hotelId",
                        "in": "query",
                        "description": "hotel Id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "parameter": "nameRoom",
                        "name": "nameRoom",
                        "in": "query",
                        "description": "name's room",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "parameter": "description",
                        "name": "description",
                        "in": "query",
                        "description": "description about room",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An example endpoint"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}